% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate}
\alias{integrate}
\title{A method for numerical integration}
\usage{
integrate(
  f,
  lower,
  upper,
  ...,
  max_subdivisions = 100L,
  relative_accuracy = .Machine$double.eps^0.25,
  absolute_accuracy = relative_accuracy,
  work_size = 4 * max_subdivisions,
  stop.on.error = TRUE
)
}
\arguments{
\item{f}{an \R function taking a numeric first argument and returning
    a numeric vector of the same length.  Returning a non-finite element will
    generate an error.}

\item{lower, upper}{the limits of integration.  Can be infinite.}

\item{...}{additional arguments to be passed to \code{f}.}

\item{max_subdivisions}{the maximum number of subintervals.}

\item{relative_accuracy}{relative accuracy requested.}

\item{absolute_accuracy}{absolute accuracy requested.}

\item{stop.on.error}{logical. If true (the default) an error stops the
    function.  If false some errors will give a result with a warning in
    the \code{message} component.}
}
\value{
A list of class \code{integrate} with components \code{value}, \code{abs.error},
}
\description{
A method for numerical integration
}
\seealso{
Other test-helper: 
\code{\link{Integrator-class}},
\code{\link{catch_what}()}
}
\concept{test-helper}
\keyword{internal}
