% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_integratecpp.R
\name{use_integratecpp}
\alias{use_integratecpp}
\title{Link to \code{integratecpp}}
\usage{
use_integratecpp(name = NULL, min_version = NULL)
}
\arguments{
\item{name}{If supplied, creates and opens ⁠src/integrate.cpp.⁠}

\item{min_version}{Optionally, supply a minimum version for the package.
Set to \code{TRUE} to use the currently installed version.}
}
\description{
Sets you package up for linking to \code{integratecpp} headers.
}
\examples{
\dontrun{
use_integratecpp()
use_integratecpp("gamma_inc")
use_integratecpp(min_version = TRUE)
}

}
