% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_what.R
\name{catch_what}
\alias{catch_what}
\title{Throw exception and catch error message}
\usage{
catch_what(exception, message)
}
\arguments{
\item{exception}{name of the exception to be thrown.}

\item{message}{error message passed to the exception constructor.}
}
\description{
Throw exception and catch error message
}
\keyword{internal}
