% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratecpp-package.R
\docType{package}
\name{integratecpp-package}
\alias{integratecpp}
\alias{integratecpp-package}
\title{integratecpp: C++ API for R Integration Routines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a header-only C++11 interface to R's C-API for numerical integration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hsloot.github.io/integratecpp/}
}

}
\author{
\strong{Maintainer}: Henrik Sloot \email{henrik.sloot@gmail.com} (\href{https://orcid.org/0000-0001-8582-7776}{ORCID})

}
