% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\docType{class}
\name{Integrator-class}
\alias{Integrator-class}
\alias{Integrator}
\alias{initialize,Integrator-method}
\alias{$,Integrator-method}
\alias{$<-,Integrator-method}
\title{A class for numerical integration}
\usage{
\S4method{initialize}{Integrator}(
  .Object,
  max_subdivisions = 100,
  relative_accuracy = .Machine$double.eps^0.25,
  absolute_accuracy = relative_accuracy,
  work_size = 4 * max_subdivisions
)

\S4method{$}{Integrator}(x, name)

\S4method{$}{Integrator}(x, name) <- value
}
\description{
A class for numerical integration
}
\section{Functions}{
\itemize{
\item \code{initialize,Integrator-method}: Construct an object of class \code{Integrator}.

\item \code{$,Integrator-method}: Either access configuration parameters
\code{max_subdivisions}, \code{relative_accuracy}, \code{absolute_accuracy}, or
\code{work_size} or get the integration routine with signature
\verb{function(f, lower, upper, ..., stop.on.error = TRUE)}.

\item \code{$<-,Integrator-method}: Set any of the configuration parameters
\code{max_subdivisions}, \code{relative_accuracy}, \code{absolute_accuracy},
or \code{work_size}.
}}

\section{Slots}{

\describe{
\item{\code{pointer}}{An external pointer to a C++ object.}
}}

\keyword{internal}
